pyproject.toml
[project]
name = "bookshelf-api"
version = "0.1.0"
description = "A RESTful API for tracking your reading list, built with FastAPI and SQLite"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.0",
    "fastapi>=0.135.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.24",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]
